Added support for packaging with Briefcase.
Running 'python setup.py macos' will now convert the code
into a standalone macOS .app.

This may not correctly install all the dependencies (due to
them not being on PyPI); however, if you pip install them
into macOS/Myriagon.app/Contents/Resources/app_packages,
the app should start.
